+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/main/DoEverythingWarmupHandler.scala
@@ -1,29 +1,37 @@
 package com.twitter.finatra.http.tests.integration.doeverything.main
 
+import com.twitter.finatra.http.response.ResponseUtils._
 import com.twitter.finatra.http.routing.HttpWarmup
 import com.twitter.finatra.httpclient.RequestBuilder._
 import com.twitter.inject.Logging
 import com.twitter.inject.utils.Handler
-import javax.inject.{Inject, Singleton}
+import javax.inject.Inject
+import scala.util.control.NonFatal
 
-@Singleton
 class DoEverythingWarmupHandler @Inject()(warmup: HttpWarmup) extends Handler with Logging {
 
   override def handle(): Unit = {
     try {
-      warmup.send(get("/ok"))
+      warmup.send(get("/ok"))()
 
-      warmup.send(post("/post"))
+      warmup.send(post("/post"))(expectOkResponse)
 
-      warmup.send(put("/put"))
+      warmup.send(put("/put"))()
 
-      warmup.send(delete("/delete"))
+      warmup.send(delete("/delete"))()
 
-      warmup.send(get("/admin/finatra/foo"))
+      /* user-defined admin route should be available in warmup */
+      warmup.send(get("/admin/testme"), admin = true)(expectOkResponse)
 
-      warmup.send(get("/health"), forceRouteToAdminHttpMuxers = true)
+      // should not have to specify "admin = true" here.
+      warmup.send(get("/admin/finatra/internal/route"))(expectOkResponse)
+
+      /* TwitterServer HTTP Admin Interface admin routes are not routable via the HttpWarmup utility */
+      warmup.send(get("/health"), admin = true)(expectNotFoundResponse)
     } catch {
-      case e: Throwable =>
+      case e: java.lang.AssertionError =>
+        throw e // any assertion error will fail the server startup
+      case NonFatal(e) =>
         // we don't want failed warmup to prevent the server from starting
         error(e.getMessage, e)
     }