+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/test/DoEverythingServerFeatureTest.scala
@@ -2445,5 +2445,36 @@ class DoEverythingServerFeatureTest extends FeatureTest with Mockito {
       "/httpclient",
       andExpect = Status.InternalServerError)
   }
+
+  test("DoEverythingServer#support specifying GlobalFlags") {
+    var shouldLogMetrics = false
+
+    com.twitter.finagle.stats.logOnShutdown() should equal(false) // verify initial default value
+
+    com.twitter.finagle.stats.logOnShutdown.let(false) { //set the scope of this test thread
+      val srvr = new EmbeddedHttpServer(
+        twitterServer = new DoEverythingServer {
+          override protected def postInjectorStartup(): Unit = {
+            // mutate to match the inner scope of withLocals
+            shouldLogMetrics = com.twitter.finagle.stats.logOnShutdown()
+            super.postInjectorStartup()
+          }
+        },
+        globalFlags = Map(
+          com.twitter.finagle.stats.logOnShutdown -> "true"
+        )
+      )
+      try {
+        srvr.start() // start the server, otherwise the scope will never be entered
+        shouldLogMetrics should equal(true) // verify mutation of inner scope
+        com.twitter.finagle.stats
+          .logOnShutdown() should equal(false) // verify outer scope is not changed
+      } finally {
+        srvr.close()
+      }
+    }
+
+    com.twitter.finagle.stats.logOnShutdown() should equal(false) // verify default value unchanged
+  }
 }