+++ inject/inject-utils/src/main/scala/com/twitter/inject/utils/Handler.scala
@@ -1,5 +1,7 @@
 package com.twitter.inject.utils
 
 abstract class Handler {
+
+  /** Executes the function of this handler. **/
   def handle(): Unit
 }